+++ src/main/java/com/akiba/backend/used/dto/response/UsedPostDetailResponse.java
@@ -1,10 +1,14 @@
 package com.akiba.backend.used.dto.response;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 import lombok.Getter;
+import java.math.BigDecimal;
 
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 import java.util.List;
+import java.math.BigDecimal;
 
 @Getter
 @Builder
@@ -14,7 +18,7 @@ public class UsedPostDetailResponse {
     private String type;
     private String title;
     private String content;
-    private Integer price;
+    private BigDecimal price;
     private String productCondition;
     private String specialType;
     private String status;
